+++ CHANGELOG.md
@@ -1,1 +1,3 @@
-Hotfix stable release
+### Additions
+- Add compat with rainbow's foliage: polytone (vanilla leaves)
+- Compat for pack's modded leaves to come